+++ tasks/kulik_a_mat_mul_double_ccs/omp/src/ops_omp.cpp
@@ -11,7 +11,7 @@
 
 namespace kulik_a_mat_mul_double_ccs {
 
-void KulikAMatMulDoubleCcsOMP::ProcessColumn(int j, int tid, const CCS &a, const CCS &b,
+void KulikAMatMulDoubleCcsOMP::ProcessColumn(size_t j, int tid, const CCS &a, const CCS &b,
                                              std::vector<std::vector<double>> &thread_accum,
                                              std::vector<std::vector<bool>> &thread_nz,
                                              std::vector<std::vector<size_t>> &thread_nnz_rows,
